RDCC-872 renamed Test Data Builders, removing CVEs
