+++ src/test/java/uk/gov/hmcts/reform/userprofileapi/controller/response/UserProfileResponseTest.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.userprofileapi.controller.response;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileDataTestBuilder.buildCreateUserProfileData;
+import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileTestDataBuilder.buildCreateUserProfileData;
 
 import org.junit.Test;
 import org.springframework.http.HttpStatus;
