+++ src/test/java/uk/gov/hmcts/reform/userprofileapi/controller/response/UserProfileWithRolesResponseTest.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.userprofileapi.controller.response;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileDataTestBuilder.buildCreateUserProfileData;
+import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileTestDataBuilder.buildCreateUserProfileData;
 
 import java.util.ArrayList;
 import java.util.List;
