+++ src/test/java/uk/gov/hmcts/reform/userprofileapi/helper/UserProfileTestDataBuilder.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.userprofileapi.helper;
 
 import static org.apache.commons.lang.RandomStringUtils.randomAlphanumeric;
-import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileDataTestBuilder.buildCreateUserProfileData;
+import static uk.gov.hmcts.reform.userprofileapi.helper.CreateUserProfileTestDataBuilder.buildCreateUserProfileData;
 
 import java.lang.reflect.Field;
 import java.time.LocalDateTime;
